Fix #31: initial click command line
