+++ emu/__init__.py
@@ -1,4 +1,4 @@
 from .wsgi import application
-from .demo import main
+from .cli import cli
 
 __version__ = "0.8.0"
