+++ emu/templates/pywps.cfg
@@ -2,12 +2,12 @@
 url = http://{{ wps_hostname }}:{{ wps_port }}/wps
 outputurl = http://{{ wps_hostname }}:{{ wps_port }}/outputs
 allowedinputpaths = /
-maxsingleinputsize = {{ maxsingleinputsize }}
-maxprocesses = {{ wps_maxprocesses }}
-parallelprocesses = {{ wps_parallelprocesses }}
+maxsingleinputsize = {{ wps_maxsingleinputsize|default('200mb') }}
+maxprocesses = {{ wps_maxprocesses|default('10') }}
+parallelprocesses = {{ wps_parallelprocesses|default('2') }}
 
 [logging]
-level = {{ wps_log_level }}
-file = {{ wps_log_file }}
-database = {{ wps_database }}
+level = {{ wps_log_level|default('INFO') }}
+file = {{ wps_log_file|default('pywps.log') }}
+database = {{ wps_database|default('sqlite:///pywps-logs.sqlite') }}
 format = %(asctime)s] [%(levelname)s] line=%(lineno)s module=%(module)s %(message)s
