+++ requirements.txt
@@ -1,2 +1,3 @@
 pywps>=4.0.0
 jinja2
+click
