+++ setup.cfg
@@ -17,6 +17,7 @@ exclude =
     docs/source/conf.py,
     build,
     dist,
+    src,
 
 [doc8]
 ignore-path = docs/build,docs/source/_templates,docs/source/_static
